refactor(auth): rename require-auth to requireAuth, fix hoisting bug and prisma config
