+++ prisma.config.ts
@@ -1,5 +1,3 @@
-// This file was generated by Prisma, and assumes you have installed the following:
-// npm install --save-dev prisma dotenv
 import { config } from "dotenv";
 
 config({ path: ".env.local" });
@@ -10,8 +8,5 @@ export default defineConfig({
   schema: "prisma/schema.prisma",
   migrations: {
     path: "prisma/migrations",
-  },
-  datasource: {
-    url: process.env["DATABASE_URL"],
   },
 });
